Backend Code/Model Training/requirements.txt
Flask==1.1.2
google-cloud-vision==2.0.0
google-cloud-automl==2.0.0
pillow==7.2.0
numpy==1.19.2
opencv-python==4.4.0.44
opencv-contrib-python==4.4.0.44
